requirement.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-pinecone
pinecone-client
python-dotenv
pymupdf
